requirements.txt
langchain==0.0.180
chromadb==0.3.25
faiss-cpu>=1.7.3
openai>=0.26.2
docx2txt>=0.8
pillow>=9.4.0
pypdf
EbookLib
tenacity>=8.2.0
path>=16.6.0
tiktoken>=0.4.0
transformers>=4.30
python-dotenv>=0.21.1
pycryptodome
bs4

